Changes for redis and docker test to run
